Modularizes text animation and background creation
Extracts text animation logic and background element creation into separate modules for improved code organization and reusability. Updates main script to initialize these features after the loading sequence.
